demos/image_classification_example/requirements.txt
numpy==1.23.5
grpcio==1.53.0
tensorflow==2.11.1
tensorflow-serving-api==2.11.1
opencv-python-headless==4.7.0.72
paho-mqtt==1.6.1
bentoml==1.0.20